Refactor: delete 로직

+++ server/src/main/java/com/ryc/api/v2/role/service/ClubRoleService.java
@@ -2,6 +2,7 @@ package com.ryc.api.v2.role.service;
 
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 import org.springframework.context.event.EventListener;
@@ -10,9 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.ryc.api.v2.admin.domain.Admin;
 import com.ryc.api.v2.admin.domain.event.AdminDeletedEvent;
-import com.ryc.api.v2.admin.service.AdminService;
 import com.ryc.api.v2.club.domain.Club;
-import com.ryc.api.v2.club.domain.ClubRepository;
 import com.ryc.api.v2.club.domain.event.ClubDeletedEvent;
 import com.ryc.api.v2.club.presentation.dto.response.DetailClubResponse;
 import com.ryc.api.v2.common.dto.response.FileGetResponse;
@@ -24,8 +23,7 @@ import com.ryc.api.v2.file.service.FileService;
 import com.ryc.api.v2.role.domain.ClubRoleRepository;
 import com.ryc.api.v2.role.domain.enums.Role;
 import com.ryc.api.v2.role.domain.vo.ClubRole;
-import com.ryc.api.v2.role.presentation.dto.response.AdminsGetResponse;
-import com.ryc.api.v2.role.presentation.dto.response.RoleDemandResponse;
+import com.ryc.api.v2.role.presentation.dto.response.ClubRoleGetResponse;
 
 import lombok.RequiredArgsConstructor;
 
@@ -34,18 +32,7 @@ import lombok.RequiredArgsConstructor;
 public class ClubRoleService {
 
   private final ClubRoleRepository clubRoleRepository;
-  private final ClubRepository clubRepository;
-  private final AdminService adminService;
   private final FileService fileService;
-
-  @Transactional
-  public RoleDemandResponse assignRole(String userId, String clubId) {
-    Club club = clubRepository.findById(clubId);
-    Admin admin = adminService.getAdminById(userId);
-    ClubRole clubRole = assignRole(admin, club, Role.MEMBER);
-
-    return new RoleDemandResponse(clubRole.id());
-  }
 
   @Transactional
   public ClubRole assignRole(Admin admin, Club club, Role role) {
@@ -54,19 +41,29 @@ public class ClubRoleService {
   }
 
   @Transactional(readOnly = true)
-  public List<AdminsGetResponse> getAdminsInClub(String clubId) {
+  public List<ClubRoleGetResponse> getClubRoles(String clubId) {
     List<ClubRole> clubRoles = clubRoleRepository.findRolesByClubId(clubId);
+
+    // TODO: clubid -> adminId 변경 필요,
+    //  CLUB_PROFILE -> ADMIN_PROFILE 변경 필요
+    FileGetResponse representativeImage =
+        fileService.findAllByAssociatedId(clubId).stream()
+            .filter(image -> image.getFileDomainType() == FileDomainType.CLUB_PROFILE)
+            .findFirst()
+            .map(image -> FileGetResponse.of(image, fileService.getPublicFileGetUrl(image)))
+            .orElse(null);
 
     return clubRoles.stream()
         .map(
-            clubRole ->
-                AdminsGetResponse.builder()
-                    .adminId(clubRole.admin().getId())
-                    .name(clubRole.admin().getName())
-                    .imageUrl(clubRole.admin().getImageUrl())
-                    .thumbnailUrl(clubRole.admin().getThumbnailUrl())
-                    .role(clubRole.role().toString())
-                    .build())
+            clubRole -> {
+              Admin admin = clubRole.getAdmin();
+              return ClubRoleGetResponse.builder()
+                  .adminId(admin.getId())
+                  .adminName(admin.getName())
+                  .role(clubRole.getRole().toString())
+                  .fileGetResponse(representativeImage)
+                  .build();
+            })
         .toList();
   }
 
@@ -76,11 +73,12 @@ public class ClubRoleService {
       throw new ClubException(ClubErrorCode.CLUB_OWNER_CANNOT_BE_DELETED);
     }
 
-    if (!clubRoleRepository.existsByAdminIdAndClubId(targetUserId, clubId)) {
-      throw new ClubException(ClubErrorCode.CLUB_MEMBER_NOT_FOUND);
+    if (!clubRoleRepository.existsByAdminIdAndClubId(adminId, clubId)) {
+      throw new NoSuchElementException(
+          "Club roles not found for adminId: " + adminId + " and clubId: " + clubId);
     }
 
-    clubRoleRepository.deleteByUserId(targetUserId);
+    clubRoleRepository.deleteByAdminIdAndClubId(targetUserId, clubId);
   }
 
   @Transactional(readOnly = true)
@@ -137,12 +135,20 @@ public class ClubRoleService {
   @EventListener
   @Transactional
   protected void handleClubDeletedEvent(ClubDeletedEvent event) {
+    if (!clubRoleRepository.existsByClubId(event.clubId())) {
+      return;
+    }
+
     clubRoleRepository.deleteByClubId(event.clubId());
   }
 
   @Transactional
   @EventListener
   protected void handleAdminDeletedEvent(AdminDeletedEvent event) {
-    clubRoleRepository.deleteByUserId(event.adminId());
+    if (!clubRoleRepository.existsByAdminId(event.adminId())) {
+      return;
+    }
+
+    clubRoleRepository.deleteAllByAdminId(event.adminId());
   }
 }
